resolving conflict + moving example changes to different pr
