+++ example/package.json
@@ -1,6 +1,6 @@
 {
   "name": "example",
-  "version": "0.2.0",
+  "version": "0.1.0",
   "private": true,
   "dependencies": {
     "react-terminal": "^1.3.1",
